SonaJobs/requirements.txt
# Master requirements - aggregated for the whole Jobs Platform
--only-binary=:all:
# It includes other requirement sets so you can run a single install command.
-r requirements-dev.txt
-r requirements-prod.txt
-r requirements-pillow.txt

# Core requirements (kept for direct reference / CI)

Django==4.2.7
psycopg2-binary==2.9.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==11.3.0
celery==5.3.4
redis==5.0.1
stripe==7.8.0
sendgrid==6.10.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
django-extensions==3.2.3
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
django-allauth==0.57.0
django-widget-tweaks==1.5.0
djangorestframework-simplejwt==5.3.0
dj-database-url==2.1.0
faker==22.6.0
pdfplumber==0.10.3
spacy==3.8.7
scikit-learn==1.3.2
numpy==1.24.3